Xcos MVC: handle graphics and model port creation
Port can be created on : model.in and graphics.pin setters. This commit
share all logic on port_management.hxx and allow future use. A
templatized implementation has been used to reduce the size of the code.

+++ scilab/modules/scicos/sci_gateway/cpp/sci_scicos_new.cpp
@@ -99,7 +99,7 @@ types::Function::ReturnValue sci_scicos_new(types::typed_list &in, int _iRetCoun
         return types::Function::Error;
     }
 
-    if (type_name->getCols() != in.size())
+    if (type_name->getCols() != (int) in.size())
     {
         Scierror(999, _("%s: Wrong number of input argument: %d expected.\n"), funame.data(), type_name->getCols());
         return types::Function::Error;
